+++ packages/core-sdk/src/types/resources/license.ts
@@ -3,7 +3,7 @@ import { Address, TransactionReceipt } from "viem";
 import { WithTxOptions, TxOptions } from "../options";
 import { EncodedTxData } from "../../abi/generated";
 import { LicensingConfig } from "../common";
-import { ERC20Options } from "../utils/wip";
+import { ERC20Options } from "../options";
 
 export type LicenseApiResponse = {
   data: License;
