--- packages/core-sdk/src/types/utils/erc20.ts
@@ -1,7 +1,0 @@
-import { Address, PublicClient } from "viem";
-
-export type GetERC20BalanceParams = {
-  tokenAddress: Address;
-  walletAddress: Address;
-  rcpClient: PublicClient;
-};
